+++ packages/web-client/src/controllers/listItemEditorDialog.ts
@@ -1192,7 +1192,7 @@ export class ListItemEditorDialog {
             setReviewValue(container, value);
           },
           customFieldsGrid,
-          input,
+          input ?? undefined,
           replaceDisplay,
         );
         if (isMarkdownField) {
@@ -1275,7 +1275,9 @@ export class ListItemEditorDialog {
       attachReviewListeners(insertAtTopCheckbox);
     }
     for (const field of customFieldsSection.fields) {
-      attachReviewListeners(field.input);
+      if (field.input) {
+        attachReviewListeners(field.input);
+      }
     }
 
     const updateModeToggle = (): void => {